Changed recipes tutorial into poetry website.
I am doing the Odin Project in order to learn web development, and this was supposed to be a recipes website. I love eating, but I hate cooking, so I decided to turn this into my own small archive of poetry translations (from diverse languages into Portuguese) that I did in the past.